Extract AppearanceResolver and unify overlay window theme handling
Overlay windows (RecordingIndicator, LaunchSplash, HotkeyToggleSplash)
were hardcoding UserDefaults AppleInterfaceStyle reads, ignoring the
user's in-app Light/Dark/System preference. They were also excluded
from updateAllWindowAppearances(), so they wouldn't react to runtime
system theme changes.

- Extract appearance logic into shared AppearanceResolver enum
- Overlay controllers now use AppearanceResolver.resolved() which
  respects Settings.shared.appearanceTheme
- Add updateAppearance() to all three overlay controllers
- Include overlay windows in updateAllWindowAppearances()
