+++ Sources/LookMaNoHands/Views/HotkeyToggleSplashWindowController.swift
@@ -41,9 +41,8 @@ class HotkeyToggleSplashWindowController {
         window.hasShadow = false
         window.isMovable = false
 
-        // Set concrete appearance to avoid stale inheritance under hardened runtime
-        let isDark = UserDefaults.standard.string(forKey: "AppleInterfaceStyle") == "Dark"
-        window.appearance = NSAppearance(named: isDark ? .darkAqua : .aqua)
+        // Set concrete appearance respecting user's in-app theme preference
+        window.appearance = AppearanceResolver.resolved()
 
         // Position in upper third of screen
         positionWindow(window)
@@ -152,6 +151,11 @@ class HotkeyToggleSplashWindowController {
             self.window = nil
             NSLog("✅ HotkeyToggleSplash: Dismissed (no animation)")
         }
+    }
+
+    /// Update window appearance to match current theme setting.
+    func updateAppearance() {
+        window?.appearance = AppearanceResolver.resolved()
     }
 
     /// Public method to manually dismiss (called if needed)
